+++ hcloud/schema.go
@@ -712,8 +712,8 @@ func loadBalancerCreateOptsToSchema(opts LoadBalancerCreateOpts) schema.LoadBala
 			DestinationPort: service.DestinationPort,
 			Proxyprotocol:   service.Proxyprotocol,
 		}
-		switch service.Protocol {
-		case LoadBalancerServiceProtocolHTTP, LoadBalancerServiceProtocolHTTPS:
+
+		if service.HTTP != nil {
 			schemaService.HTTP = &schema.LoadBalancerCreateRequestServiceHTTP{
 				RedirectHTTP:   service.HTTP.RedirectHTTP,
 				StickySessions: service.HTTP.StickySessions,
@@ -730,6 +730,7 @@ func loadBalancerCreateOptsToSchema(opts LoadBalancerCreateOpts) schema.LoadBala
 				schemaService.HTTP.Certificates = &certificates
 			}
 		}
+
 		if service.HealthCheck != nil {
 			schemaHealthCheck := &schema.LoadBalancerCreateRequestServiceHealthCheck{
 				Protocol: string(service.HealthCheck.Protocol),